+++ src/main/java/eu/printingin3d/physics/Volume.java
@@ -1,8 +1,10 @@
 package eu.printingin3d.physics;
 
+import java.math.BigDecimal;
+
 public class Volume extends DoubleValue {
 	
-	public Volume(double volume) {
+	public Volume(BigDecimal volume) {
 		super(volume);
 	}
 }
